backend/requirements.txt
fastapi
uvicorn
aiohttp
opencv-python
pinecone
python-multipart
python-dotenv
groq